Test FormListDownloader instead of task
