+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyComponent.java
@@ -26,7 +26,7 @@ import org.odk.collect.android.tasks.sms.SmsSentBroadcastReceiver;
 import org.odk.collect.android.tasks.sms.SmsService;
 import org.odk.collect.android.tasks.sms.contracts.SmsSubmissionManagerContract;
 import org.odk.collect.android.utilities.AuthDialogUtility;
-import org.odk.collect.android.utilities.DownloadFormListUtils;
+import org.odk.collect.android.utilities.FormListDownloader;
 import org.odk.collect.android.utilities.FormDownloader;
 import org.odk.collect.android.views.ODKView;
 import org.odk.collect.android.widgets.ExStringWidget;
@@ -125,7 +125,7 @@ public interface AppDependencyComponent {
 
     OpenRosaHttpInterface openRosaHttpInterface();
 
-    DownloadFormListUtils downloadFormListUtils();
+    FormListDownloader downloadFormListUtils();
 
     ReferenceManager referenceManager();
